Refactoring: put unit tests in the proper files
